refactoring players
