+++ src/components/scrimmage/ScrimmageApp.vue
@@ -36,7 +36,7 @@ import Vue from 'vue'
 import moment from 'moment'
 import Timer from './Timer'
 import Court from './Court'
-import Players from './Players'
+import Players from './players/Players'
 import Scoreboard from './score-view/Scoreboard'
 import Actions from './actions/Actions'
 import Events from './events/Events'
