+++ src/components/scrimmage/players/PlayerCard.vue
